ext/dv/: Fix seeking in dvdemux again, add some more debug info.
Original commit message from CVS:
* ext/dv/gstdvdec.c: (gst_dvdec_chain):
* ext/dv/gstdvdemux.c: (gst_dvdemux_demux_frame):
* ext/dv/gstdvdemux.h:
Fix seeking in dvdemux again, add some more debug info.

+++ ext/dv/gstdvdemux.h
@@ -73,7 +73,7 @@ struct _GstDVDemux {
   
   guint64	 timestamp;
   guint64	 duration;
-  guint          total_frames;
+  gint64         total_frames;
   guint64	 audio_offset;
   guint64	 video_offset;
